Add CICD workflow for backend Docker container (#27)
* Add CICD workflow for backend Docker container

This GitHub Actions workflow automates building and pushing the backend Docker image on changes to the main branch. It supports multi-platform builds and uses caching for improved efficiency. The workflow also securely authenticates with Docker Hub using stored secrets.

* Deleted .github from wrong dir

* Add JDK setup and Maven build step to CI workflow

Updated the backend CI workflow to set up JDK 21 using GraalVM and added a Maven build step with necessary environment variables. These changes ensure the project compiles and packages properly before proceeding with Docker operations.

* Update GitHub Actions trigger for backend CICD workflow

Removed pull_request trigger to streamline the workflow, previous intention was just for test purpose

* Enable Docker image push in CI/CD workflow

The `push` parameter was set to `true` in the Docker build step, ensuring that built images are now pushed to the specified Docker registry. This change is critical for seamless deployment in the CI/CD pipeline.

* Remove version declaration from docker-compose.yaml

The 'version' field is no longer required in recent Docker Compose versions. This change ensures compatibility with the latest standards.
```
docker-compose.yaml: the attribute `version` is obsolete, it will be ignored, please remove it to avoid potential confusion
```